+++ src/main.py
@@ -2,18 +2,27 @@ import cli.main as main_cli
 from server import main
 import sys
 
-helpText = 'Usage: katana [command] -c [config path]'
+
+def help() -> str:
+    help_text = 'Usage: katana [command] -c [config path]\n'
+    help_text += 'Commands:\n'
+    help_text += '\t<none> - run interactive cli\n'
+    help_text += '\tstart - run http server\n'
+    help_text += '\ttests - run unit-tests\n'
+    return help_text
+
+
 config_path = ''
 
 if len(sys.argv) == 1:
-    print(helpText)
+    print(help())
     raise SystemExit
 
 if sys.argv[1] == '-c':
     if len(sys.argv) > 2:
         config_path = sys.argv[2]
     else:
-        print(helpText)
+        print(help())
         raise SystemExit
 else:
     if sys.argv[1] == 'start':
@@ -21,16 +30,16 @@ else:
             if sys.argv[2] == '-c':
                 config_path = sys.argv[3]
             else:
-                print(helpText)
+                print(help())
                 raise SystemExit
         else:
-            print(helpText)
+            print(help())
             raise SystemExit
         server = main.Server(config_path)
         server.init()
         raise SystemExit
     else:
-        print(helpText)
+        print(help())
         raise SystemExit
 
 main_cli.main(config_path)